Update README.md
This is govdigital's contribution to layer governance. The text describes how the governance aspects are regulated within govdigital. The questions from the individual RFCs are answered as far as possible for govdigital at the current time. The text refers to the contents of the following RFCs: 0003 Member Eligibility Criteria, 0004: Suspension and Termination of members, RFC 0018: Application process for new members, RFC-0019 Onboarding process for new members, What happens if responsible individuals leave a member organisation?